requirements.txt
nextcord
flask
Pillow
python-dotenv
decorator
colorama
matplotlib
PyYAML